Fix deadlock issue when on the server side:
Currently, clients are added to the list before the setup is complete. However, we implictly assume that each connection in the client list is already set up. To fix this, we move the connection accepting part to before we create the client struct. This means that once the client is added to the lit, its setup is guarenteed to happen in a finite amount of time.